pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "plancheck"
version = "0.1.0"
description = "Advanced architectural/engineering plan parser with OCR reconciliation"
requires-python = ">=3.10"
dependencies = [
    "Pillow>=12.0.0",
    "pdfplumber>=0.11.0",
    "numpy>=1.24.0",
    "opencv-python>=4.8.0",
    "pydantic>=2.7.0",
    "reportlab>=4.0.0",
    "scikit-learn>=1.4.0",
    "joblib>=1.3.0",
    "shapely>=2.0.0",
]

[project.optional-dependencies]
ocr = [
    "paddlepaddle>=3.0.0,<3.3; python_version<'3.13'",
    "paddleocr>=3.0.0; python_version<'3.13'",
]
vision = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "timm>=1.0.0",
]
layout = [
    "transformers>=4.30.0",
    "torch>=2.0.0",
]
embeddings = [
    "sentence-transformers>=2.2.0",
]
llm = [
    "ollama>=0.4.0,<1.0.0",
    "openai>=1.0.0,<2.0.0",
    "anthropic>=0.30.0,<1.0.0",
]
query = [
    "chromadb>=0.5.0",
    "sentence-transformers>=2.2.0",
]
gnn = [
    "torch>=2.0.0",
    "torch-geometric>=2.4.0",
]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
    "pytest-timeout>=2.0.0",
    "pytest-xdist>=3.0.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
timeout = 60
markers = [
    "unit: fast isolated unit tests",
    "integration: tests that combine multiple modules",
    "slow: long-running tests (>10s)",
    "gpu: tests requiring CUDA/GPU",
]
